[refactor/#179] signupSchedule test 수정
