+++ resource-server/src/test/java/com/inhabas/api/web/SignUpScheduleControllerTest.java
@@ -3,7 +3,7 @@ package com.inhabas.api.web;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.inhabas.api.domain.signUpSchedule.InvalidDateException;
 import com.inhabas.api.domain.signUpSchedule.dto.SignUpScheduleDto;
-import com.inhabas.api.domain.signUpSchedule.domain.SignUpScheduler;
+import com.inhabas.api.domain.signUpSchedule.domain.usecase.SignUpScheduler;
 import com.inhabas.testAnnotataion.NoSecureWebMvcTest;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
@@ -40,7 +40,7 @@ public class SignUpScheduleControllerTest {
     @Test
     public void getSignUpSchedule() throws Exception {
         //given
-        SignUpScheduleDto existingSchedule = new SignUpScheduleDto(1, 999,
+        SignUpScheduleDto existingSchedule = new SignUpScheduleDto(999,
                 LocalDateTime.of(2022, 1, 1, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 2, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 3, 0, 0, 0),
@@ -56,7 +56,7 @@ public class SignUpScheduleControllerTest {
     @DisplayName("회원가입 일정을 수정한다.")
     @Test
     public void updateSignUpSchedule() throws Exception {
-        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(1, 999,
+        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(999,
                 LocalDateTime.of(2022, 1, 1, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 2, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 3, 0, 0, 0),
@@ -72,7 +72,7 @@ public class SignUpScheduleControllerTest {
     @DisplayName("회원 기수가 양의 정수가 아닌경우, 400")
     @Test
     public void invalidGenerationTest() throws Exception {
-        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(1, 0,
+        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(0,
                 LocalDateTime.of(2022, 1, 1, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 2, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 3, 0, 0, 0),
@@ -93,7 +93,7 @@ public class SignUpScheduleControllerTest {
     @Test
     public void invalidDateInfoTest() throws Exception {
         doThrow(InvalidDateException.class).when(signUpScheduler).updateSchedule(any());
-        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(1, 999,
+        String jsonRequest = objectMapper.writeValueAsString(new SignUpScheduleDto(999,
                 LocalDateTime.of(2022, 1, 5, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 4, 0, 0, 0),
                 LocalDateTime.of(2022, 1, 3, 0, 0, 0),
